\name{ompbn-package}
\alias{ompbn-package}
\docType{package}
\title{
Bayesian Network Learning Using the OMP Algorithm
}

\description{
The original version of this package was to learn Bayesian networks with the PCHC algorithm. PCHC stands for PC Hill-Climbing. It is a new hybrid algorithm that used PC to construct the skeleton of the BN and then utilizes the Hill-Climbing greedy search. The package has been expanded to include the MMHC and the FEDHC algortihms. It further includes the PCTABU, MMTABU and FEDTABU algortihms which are pretty much similar. Instead of the Hill Climbing greey search the Tabu search is employed.
}

\details{
\tabular{ll}{
Package: \tab pchc\cr
Type: \tab Package\cr
Version: \tab 1.0 \cr
Date: \tab 2026-09-17\cr
License: \tab GPL-2\cr
}
}

\author{
Michail Tsagris \email{mtsagris@uoc.gr}.
}

\section{Maintainers}{
Michail Tsagris <mtsagris@uoc.gr>.
}

\references{
Sevinc V. and Tsagris M. (2024). On the Hyperparameters of PCTABU and PCHC Bayesian Network Learning Algorithms.
https://assets-eu.researchsquare.com/files/rs-5137132/v1_covered_442bc8e8-2d63-4f92-be2d-1ecf9d7f9ee8.pdf?c=1727238437

Tsagris M. (2022). The FEDHC Bayesian Network Learning Algorithm. Mathematics 2022, 10(15): 2604.

Tsagris M. (2021). A new scalable Bayesian network learning algorithm with applications to economics. Computational Economics, 57(1): 341--367.

Spirtes P.,  Glymour C. and Scheines R. (2001). Causation, Prediction, and Search.
The MIT Press, Cambridge, MA, USA, 3nd edition.

Tsamardinos I., Borboudakis G. (2010) Permutation Testing Improves Bayesian Network Learning.
In Machine Learning and Knowledge Discovery in Databases. ECML PKDD 2010. 322--337.

Tsamardinos I., Brown E.L. and Aliferis F.C. (2006). The max-min hill-climbing Bayesian network structure learning algorithm. Machine learning 65(1): 31--78.

Tsagris M. (2017). Conditional independence test for categorical data using Poisson log-linear model. Journal of Data Science, 15(2): 347--356.

Borboudakis G. and Tsamardinos I. (2019). Forward-backward selection with early dropping.
Journal of Machine Learning Research, 20(8): 1-39.
}

%\note{
%Acknowledgments:
%}



